requirements.txt
streamlit
pdfplumber
python-docx
spacy
nltk
scikit-learn
python-Levenshtein
